Conflicts:
	examples/work/hello-world.PC_multiboot
	examples/work/hello-world.cpp
	examples/work/kernel
	examples/work/kkk
	examples/work/obj_PC_multiboot/DBStream.o
	examples/work/obj_PC_multiboot/Print.o
	examples/work/obj_PC_multiboot/Stream.o
	examples/work/obj_PC_multiboot/abstract_irq.o
	examples/work/obj_PC_multiboot/allocator.o
	examples/work/obj_PC_multiboot/boot3.o
	examples/work/obj_PC_multiboot/command_list.o
	examples/work/obj_PC_multiboot/debug_print.o
	examples/work/obj_PC_multiboot/debug_putchar.o
	examples/work/obj_PC_multiboot/delay_cpu.o
	examples/work/obj_PC_multiboot/descriptor_tables.o
	examples/work/obj_PC_multiboot/dynamicring.o
	examples/work/obj_PC_multiboot/error.o
	examples/work/obj_PC_multiboot/error_platform.o
	examples/work/obj_PC_multiboot/genoslib.o
	examples/work/obj_PC_multiboot/hello-world.o
	examples/work/obj_PC_multiboot/initenv.o
	examples/work/obj_PC_multiboot/isr.o
	examples/work/obj_PC_multiboot/keyscan.o
	examples/work/obj_PC_multiboot/pci_search.o
	examples/work/obj_PC_multiboot/queue.o
	examples/work/obj_PC_multiboot/readline.o
	examples/work/obj_PC_multiboot/regread.o
	examples/work/obj_PC_multiboot/ring.o
	examples/work/obj_PC_multiboot/rl_terminal.o
	examples/work/obj_PC_multiboot/staticring.o
	examples/work/obj_PC_multiboot/timer.o
	examples/work/obj_PC_multiboot/utility.o
	examples/work/obj_PC_multiboot/video_page.o
	platform/PC_multiboot/Makefile.PC_multiboot

+++ invariant/allocator/alloc_standart.cpp
@@ -11,13 +11,13 @@ Standart_alloc galloc;
 void * Standart_alloc::allocate (size_t size)
 {	
 	tpr("alloc global"); tprtab(); tprdec(size);
-	void* ptr=malloc(size); tprtab(); tprlnhex(ptr);
+	void* ptr=malloc(size); tprtab(); tprlnhex((size_t)ptr);
 	return ptr;
 }
 
 
 void Standart_alloc::deallocate (void * p)
 {
-	tpr ("alloc global free "); tprlnhex(p);
+	tpr ("alloc global free "); tprlnhex((size_t)p);
 	free(p);
 }
